Load data as we have it
