+++ sippy-ng/src/tests/TestDurationChart.js
@@ -1,3 +1,4 @@
+import { CircularProgress } from '@material-ui/core'
 import { Error } from '@material-ui/icons'
 import { Line } from 'react-chartjs-2'
 import { safeEncodeURIComponent } from '../helpers'
@@ -54,7 +55,7 @@ export function TestDurationChart(props) {
   }
 
   if (!isLoaded) {
-    return <p>Loading...</p>
+    return <CircularProgress color="inherit" />
   }
 
   const chart = {
